added script/code for writing out specific lipids in trajectories (fixed errors)

+++ basicrta/functions.py
@@ -392,7 +392,7 @@ def write_trajs(u, time, trajtime, indicator, residue, lipind):
         eframes = eframes[sortinds]
         tmp = [np.arange(b, e) for b, e in zip(bframes, eframes)]
         tmpL = [np.ones_like(np.arange(b, e))*l for b, e, l in zip(bframes, eframes, lipind[inds[comp]])]
-        write_frames, write_Linds = np.concatenate([*tmp]), np.concatenate([*tmpL])
+        write_frames, write_Linds = np.concatenate([*tmp]), np.concatenate([*tmpL]).astype(int)
         with mda.Writer(f"{residue}/comp{comp}_traj.xtc", len((prot+chol.residues[0].atoms).atoms)) as W:
             for i, ts in tqdm(enumerate(u.trajectory[write_frames]), desc=f"{residue}-comp{comp}", position=proc,
                               leave=False, total=len(write_frames)):
